requirements.txt
huggingface_hub
onnxruntime
tqdm
sounddevice
numpy
aiohttp
python-dotenv
requests
websockets
pyinstaller